+++ src/resources/manifest.php
@@ -2,5 +2,5 @@
 
 return [
     'name'    => 'Konekt Address Module',
-    'version' => '2.5.0'
+    'version' => '2.5.1'
 ];